Adds: api functions for all edge apps commands.
