+++ src/api/asset.rs
@@ -3,10 +3,17 @@ use crate::commands::CommandError;
 use crate::commands;
 
 use serde::{Deserialize, Serialize};
+use serde_json::Value;
 
 #[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
 pub struct AssetSignature {
     pub(crate) signature: String,
+}
+#[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
+pub struct AssetProcessingStatus {
+    pub(crate) status: String,
+    pub(crate) processing_error: String,
+    pub(crate) title: String,
 }
 
 impl Api {
@@ -22,5 +29,21 @@ impl Api {
                 app_id, revision
             ),
         )?)?)
+    }
+
+    pub fn get_processing_statuses(
+        &self,
+        app_id: &str,
+        revision: u32,
+    ) -> Result<Vec<AssetProcessingStatus>, CommandError> {
+        let response = commands::get(
+            &self.authentication,
+            &format!(
+                "v4/assets?select=status,processing_error,title&app_id=eq.{}&app_revision=eq.{}&status=neq.finished",
+                app_id, revision
+            ),
+        )?;
+
+        Ok(serde_json::from_value::<Vec<AssetProcessingStatus>>(response)?)
     }
 }
