+++ src/api/edge_app/mod.rs
@@ -1,0 +1,5 @@
+pub mod app;
+pub mod setting;
+pub mod version;
+pub mod channel;
+pub mod installation;
