+++ src/commands/edge_app/manifest.rs
@@ -10,7 +10,7 @@ use std::path::Path;
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 
-use crate::commands::edge_app::setting::{deserialize_settings, serialize_settings, Setting};
+use crate::api::edge_app::setting::{Setting, serialize_settings, deserialize_settings};
 use crate::commands::serde_utils::{
     deserialize_option_string_field, string_field_is_none_or_empty,
 };
@@ -332,7 +332,7 @@ pub fn beautify_error_message(error: &str) -> String {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::commands::edge_app::setting::SettingType;
+    use crate::api::edge_app::setting::{Setting, SettingType};
     use tempfile::tempdir;
 
     fn create_test_manifest() -> EdgeAppManifest {
