+++ src/commands/edge_app/manifest_auth.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::commands::edge_app::setting::{Setting, SettingType};
+use crate::api::edge_app::setting::{Setting, SettingType};
 
 #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
