+++ src/commands/edge_app/setting.rs
@@ -1,6 +1,8 @@
 use crate::commands;
 use crate::commands::edge_app::EdgeAppCommand;
 use crate::commands::{CommandError, EdgeAppSettings};
+use crate::api::edge_app::setting::Setting;
+
 use log::debug;
 use std::collections::HashMap;
 use std::str;
@@ -8,210 +10,11 @@ use std::str;
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 
-use std::ops::Not;
-use std::str::FromStr;
-
-use serde::Deserializer;
-use strum::IntoEnumIterator;
-use strum_macros::{Display, EnumIter, EnumString};
-
-use crate::commands::serde_utils::{deserialize_string_field, serialize_non_empty_string_field};
-
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Default, EnumString, Display, EnumIter)]
-pub enum SettingType {
-    #[default]
-    #[strum(serialize = "string")]
-    String,
-    #[strum(serialize = "secret")]
-    Secret,
-}
-
-// maybe we can use a better name as we have EdgeAppSettings which is the same but serde_json::Value inside
-#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct Setting {
-    #[serde(
-        rename = "type",
-        serialize_with = "serialize_setting_type",
-        deserialize_with = "deserialize_setting_type"
-    )]
-    pub type_: SettingType,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub default_value: Option<String>,
-    #[serde(default)]
-    pub title: Option<String>,
-    #[serde(skip)]
-    pub name: String,
-    pub optional: bool,
-
-    #[serde(
-        serialize_with = "serialize_help_text",
-        deserialize_with = "deserialize_help_text"
-    )]
-    pub help_text: String,
-
-    #[serde(default = "bool::default", skip_serializing_if = "<&bool>::not")]
-    pub is_global: bool,
-}
-
-pub fn serialize_settings<S>(settings: &[Setting], serializer: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    use serde::ser::SerializeMap;
-
-    let mut map = serializer.serialize_map(Some(settings.len()))?;
-    for setting in settings {
-        map.serialize_entry(&setting.name, &setting)?;
-    }
-    map.end()
-}
-
-pub fn deserialize_settings<'de, D>(deserializer: D) -> Result<Vec<Setting>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let map: HashMap<String, Setting> = serde::Deserialize::deserialize(deserializer)?;
-    let mut settings: Vec<Setting> = map
-        .into_iter()
-        .map(|(name, mut setting)| {
-            setting.name = name;
-            setting
-        })
-        .collect();
-
-    for setting in &settings {
-        if setting.type_ == SettingType::Secret && setting.default_value.is_some() {
-            return Err(serde::de::Error::custom(format!(
-                "Setting \"{}\" is of type \"secret\" and cannot have a default value",
-                setting.name
-            )));
-        }
-        if setting.name.starts_with("screenly_") {
-            return Err(serde::de::Error::custom(format!(
-                "Setting \"{}\" cannot start with \"screenly_\" as this prefix is preserved.",
-                setting.name
-            )));
-        }
-    }
-
-    settings.sort_by_key(|s| s.name.clone());
-    Ok(settings)
-}
-
-pub fn deserialize_settings_from_array<'de, D>(deserializer: D) -> Result<Vec<Setting>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let map: Vec<HashMap<String, Value>> = serde::Deserialize::deserialize(deserializer)?;
-    let mut settings: Vec<Setting> = map
-        .into_iter()
-        .map(|setting_data| {
-            let mut setting = Setting::default();
-            for (key, value) in setting_data {
-                match key.as_str() {
-                    "type" => {
-                        setting.type_ =
-                            deserialize_setting_type(value).expect("Failed to parse setting type.");
-                    }
-                    "default_value" => {
-                        setting.default_value = value.as_str().map(|s| s.to_string());
-                    }
-                    "title" => {
-                        setting.title = value.as_str().map(|s| s.to_string());
-                    }
-                    "optional" => {
-                        setting.optional = value.as_bool().expect("Failed to parse optional.")
-                    }
-                    "help_text" => {
-                        setting.help_text = value
-                            .as_str()
-                            .expect("Failed to parse help_text.")
-                            .to_string();
-                    }
-                    "is_global" => {
-                        setting.is_global = value.as_bool().expect("Failed to parse is_global.");
-                    }
-                    "name" => {
-                        setting.name = value.as_str().expect("Failed to parse name.").to_string();
-                    }
-                    _ => {}
-                }
-            }
-            setting
-        })
-        .collect();
-
-    settings.sort_by_key(|s| s.name.clone());
-    Ok(settings)
-}
-
-fn serialize_setting_type<S>(setting_type: &SettingType, serializer: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    serializer.serialize_str(&setting_type.to_string())
-}
-
-fn deserialize_setting_type<'de, D>(deserializer: D) -> Result<SettingType, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let s: String = Deserialize::deserialize(deserializer)?;
-
-    let valid_setting_types = SettingType::iter()
-        .map(|t| t.to_string())
-        .collect::<Vec<_>>()
-        .join("\n")
-        + "\n";
-
-    match SettingType::from_str(&s.to_lowercase()) {
-        Ok(setting_type) => Ok(setting_type),
-        Err(_) => Err(serde::de::Error::custom(format!(
-            "Setting type should be one of the following:\n{}",
-            valid_setting_types
-        ))),
-    }
-}
-
-fn serialize_help_text<S>(value: &str, serializer: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    serialize_non_empty_string_field("help_text", value, serializer)
-}
-
-fn deserialize_help_text<'de, D>(deserializer: D) -> Result<String, D::Error>
-where
-    D: serde::de::Deserializer<'de>,
-{
-    deserialize_string_field("help_text", true, deserializer)
-}
-
-impl Setting {
-    pub fn new(type_: SettingType, title: &str, name: &str, help_text: &str, global: bool) -> Self {
-        Setting {
-            type_,
-            default_value: None,
-            title: Some(title.to_string()),
-            name: name.to_string(),
-            optional: false,
-            help_text: help_text.to_string(),
-            is_global: global,
-        }
-    }
-}
 
 impl EdgeAppCommand {
-    pub fn list_settings(&self, installation_id: &str) -> Result<EdgeAppSettings, CommandError> {
-        let app_id = self.get_app_id_by_installation(installation_id)?;
-
-        let app_settings: Vec<HashMap<String, serde_json::Value>> = serde_json::from_value(commands::get(&self.api.authentication,
-                                                                                                             &format!("v4.1/edge-apps/settings?select=name,type,default_value,optional,title,help_text,edge_app_setting_values(value)&app_id=eq.{}&order=name.asc",
-                                                                                                                      app_id,
-                                                                                                             ))?)?;
-
-        Ok(EdgeAppSettings::new(serde_json::to_value(app_settings)?))
+    pub fn list_settings(&self, path: Option<String>) -> Result<EdgeAppSettings, CommandError> {
+        let app_id = self.get_app_id(path)?;
+        Ok(self.api.list_settings(&app_id)?)
     }
 
     pub fn set_setting(
@@ -224,12 +27,16 @@ impl EdgeAppCommand {
             Ok(id) => Some(id),
             Err(_) => None,
         };
+        let actual_installation_id = match installation_id {
+            Some(id) => id.clone(),
+            None => "".to_string(),
+        };
         let app_id: String = match self.get_app_id(path.clone()) {
             Ok(id) => id,
             Err(_) => return Err(CommandError::MissingAppId),
         };
 
-        let _is_setting_global = self.is_setting_global(&app_id, setting_key)?;
+        let _is_setting_global = self.api.is_setting_global(&app_id, setting_key)?;
 
         #[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
         struct SettingValue {
@@ -242,12 +49,9 @@ impl EdgeAppCommand {
         let setting_url: String;
         let settings_values_payload: Value;
         let settings_values_patch_url: String;
+        let server_setting_value;
 
         if _is_setting_global {
-            setting_url = format!(
-                "v4.1/edge-apps/settings?select=name,type,edge_app_setting_values(value)&app_id=eq.{}&edge_app_setting_values.app_id=eq.{}&name=eq.{}",
-                app_id, app_id, setting_key,
-            );
             settings_values_payload = json!(
                 {
                     "app_id": app_id,
@@ -255,20 +59,12 @@ impl EdgeAppCommand {
                     "value": setting_value,
                 }
             );
-            settings_values_patch_url = format!(
-                "v4.1/edge-apps/settings/values?app_id=eq.{}&name=eq.{}&installation_id=is.null",
-                app_id, setting_key,
-            );
+            server_setting_value = self.api.get_global_setting(&app_id, setting_key)?;
         } else {
-            let actual_installation_id = match installation_id {
-                Some(id) => id,
-                None => return Err(CommandError::MissingInstallationId),
-            };
+            if actual_installation_id == "" {
+                return Err(CommandError::MissingInstallationId);
+            }
 
-            setting_url = format!(
-                "v4.1/edge-apps/settings?select=name,type,edge_app_setting_values(value)&edge_app_setting_values.installation_id=eq.{}&name=eq.{}&app_id=eq.{}",
-                actual_installation_id, setting_key, app_id
-            );
             settings_values_payload = json!(
                 {
                     "installation_id": actual_installation_id,
@@ -276,41 +72,40 @@ impl EdgeAppCommand {
                     "value": setting_value,
                 }
             );
-            settings_values_patch_url = format!(
-                "v4.1/edge-apps/settings/values?installation_id=eq.{}&name=eq.{}",
-                actual_installation_id, setting_key,
-            );
+            server_setting_value = self.api.get_local_setting(&app_id, &actual_installation_id, setting_key)?;
+        }
+       
+        if server_setting_value.is_none() {
+            return Err(CommandError::SettingDoesNotExist(setting_key.to_string()));
         }
 
-        let response = commands::get(&self.api.authentication, &setting_url)?;
-        let setting_values = serde_json::from_value::<Vec<SettingValue>>(response)?;
-
-        if setting_values.is_empty() {
-            commands::post(
-                &self.api.authentication,
-                "v4.1/edge-apps/settings/values",
-                &settings_values_payload,
-            )?;
-            return Ok(());
-        }
         // we do know it is not empty - so it is safe to unwrap
-        let setting = setting_values.first().unwrap();
+        let setting = server_setting_value.unwrap();
 
         if setting.type_field == "secret" {
-            commands::post(
-                &self.api.authentication,
-                "v4.1/edge-apps/secrets/values",
-                &settings_values_payload,
-            )?;
+            if _is_setting_global {
+                self.api.create_global_secret_value(&app_id, setting_key, setting_value)?;
+            }
+            else {
+                if actual_installation_id == "" {
+                    return Err(CommandError::MissingInstallationId);
+                }
+                self.api.create_local_secret_value(&actual_installation_id, setting_key, setting_value)?;
+            }
+            
             return Ok(());
         }
 
         if setting.edge_app_setting_values.is_empty() {
-            commands::post(
-                &self.api.authentication,
-                "v4.1/edge-apps/settings/values",
-                &settings_values_payload,
-            )?;
+            if _is_setting_global {
+                self.api.create_global_setting_value(&app_id, setting_key, setting_value)?;
+            }
+            else {
+                if actual_installation_id == "" {
+                    return Err(CommandError::MissingInstallationId);
+                }
+                self.api.create_local_setting_value(&app_id, &actual_installation_id, setting_key, setting_value)?;
+            }
 
             return Ok(());
         }
@@ -321,33 +116,24 @@ impl EdgeAppCommand {
             println!("Setting value is already set to {}", setting_value);
             return Ok(());
         }
-        commands::patch(
-            &self.api.authentication,
-            &settings_values_patch_url,
-            &json!(
-                {
-                    "value": setting_value,
-                }
-            ),
-        )?;
+        
+        if _is_setting_global {
+            self.api.update_global_setting_value(&app_id, setting_key, setting_value)?;
+        }
+        else {
+            if actual_installation_id == "" {
+                return Err(CommandError::MissingInstallationId);
+            }
+            self.api.update_local_setting_value( &actual_installation_id, setting_key, setting_value)?;
+        }
 
         Ok(())
     }
 
     pub fn create_setting(&self, app_id: String, setting: &Setting) -> Result<(), CommandError> {
-        let value = serde_json::to_value(setting)?;
-        let mut payload = serde_json::from_value::<HashMap<String, serde_json::Value>>(value)?;
-        payload.insert("app_id".to_owned(), json!(app_id));
-        payload.insert("name".to_owned(), json!(setting.name));
-
-        debug!("Creating setting: {:?}", &payload);
-
-        let response = commands::post(&self.api.authentication, "v4.1/edge-apps/settings", &payload);
+        let response = self.api.create_setting(&app_id, setting);
         if response.is_err() {
-            let c = commands::get(
-                &self.api.authentication,
-                &format!("v4.1/edge-apps/settings?app_id=eq.{}", app_id),
-            )?;
+            let c = self.api.get_settings(&app_id)?;
             debug!("Existing settings: {:?}", c);
             return Err(CommandError::NoChangesToUpload("".to_owned()));
         }
@@ -356,21 +142,7 @@ impl EdgeAppCommand {
     }
 
     pub fn update_setting(&self, app_id: String, setting: &Setting) -> Result<(), CommandError> {
-        let value = serde_json::to_value(setting)?;
-        let mut payload = serde_json::from_value::<HashMap<String, serde_json::Value>>(value)?;
-        payload.insert("name".to_owned(), json!(setting.name));
-
-        debug!("Updating setting: {:?}", &payload);
-
-        let response = commands::patch(
-            &self.api.authentication,
-            &format!(
-                "v4.1/edge-apps/settings?app_id=eq.{id}&name=eq.{name}",
-                id = app_id,
-                name = setting.name
-            ),
-            &payload,
-        );
+        let response = self.api.update_setting(&app_id, setting);
 
         if let Err(error) = response {
             debug!("Failed to update setting: {}", setting.name);
@@ -381,14 +153,7 @@ impl EdgeAppCommand {
     }
 
     pub fn delete_setting(&self, app_id: String, setting: &Setting) -> Result<(), CommandError> {
-        let response = commands::delete(
-            &self.api.authentication,
-            &format!(
-                "v4.1/edge-apps/settings?app_id=eq.{id}&name=eq.{name}",
-                id = app_id,
-                name = setting.name
-            ),
-        );
+        let response = self.api.delete_setting(&app_id, setting);
 
         if let Err(error) = response {
             debug!("Failed to delete setting: {}", setting.name);
@@ -404,30 +169,15 @@ mod tests {
     use super::*;
     use std::env;
 
+    use crate::api::edge_app::setting::SettingType;
+
     use crate::commands::edge_app::test_utils::tests::prepare_edge_apps_test;
     use httpmock::Method::{GET, PATCH, POST};
 
     #[test]
     fn test_list_settings_should_send_correct_request() {
-        let (_tmp_dir, command, mock_server, _manifest, instance_manifest) =
-            prepare_edge_apps_test(false, true);
-
-        let installations_get_mock = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/installations")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("select", "app_id")
-                .query_param("id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEB");
-            then.status(200).json_body(json!([
-                {
-                    "app_id": "02H2QZ6Z8WXWNDC0KQ198XCZEW"
-                }
-            ]));
-        });
+        let (tmp_dir, command, mock_server, _manifest, _instance_manifest) =
+            prepare_edge_apps_test(true, false);
 
         // &format!("v4.1/edge-apps/settings?select=name,type,default_value,optional,title,help_text,edge_app_setting_values(value)&app_id=eq.{}&order=name.asc",
         let settings_mock = mock_server.mock(|when, then| {
@@ -439,7 +189,7 @@ mod tests {
                         format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
                     )
                     .query_param("select", "name,type,default_value,optional,title,help_text,edge_app_setting_values(value)")
-                    .query_param("app_id", "eq.02H2QZ6Z8WXWNDC0KQ198XCZEW")
+                    .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW")
                     .query_param("order", "name.asc");
 
                 then.status(200).json_body(json!([
@@ -490,9 +240,8 @@ mod tests {
                 ]));
             });
 
-        let result = command.list_settings(&instance_manifest.unwrap().id.unwrap());
+        let result = command.list_settings(Some(tmp_dir.path().to_str().unwrap().to_string()));
 
-        installations_get_mock.assert();
         settings_mock.assert();
 
         assert!(result.is_ok());
@@ -573,7 +322,6 @@ mod tests {
             ]));
         });
 
-        // "v4/edge-apps/settings/values?select=title&installation_id=eq.{}&title=eq.{}"
         let setting_mock_get = mock_server.mock(|when, then| {
             when.method(GET)
                 .path("/v4.1/edge-apps/settings")
@@ -589,7 +337,13 @@ mod tests {
                     "eq.01H2QZ6Z8WXWNDC0KQ198XCZEB",
                 )
                 .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW");
-            then.status(200).json_body(json!([]));
+            then.status(200).json_body(json!([
+                {
+                    "name": "best_setting",
+                    "type": "string",
+                    "edge_app_setting_values": []
+                }
+            ]));
         });
 
         let setting_values_mock_post = mock_server.mock(|when, then| {
@@ -829,7 +583,13 @@ mod tests {
                     "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW",
                 )
                 .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW");
-            then.status(200).json_body(json!([]));
+            then.status(200).json_body(json!([
+                {
+                    "name": "best_setting",
+                    "type": "string",
+                    "edge_app_setting_values": []
+                }
+            ]));
         });
 
         let setting_values_mock_post = mock_server.mock(|when, then| {
