+++ src/commands/edge_app/utils.rs
@@ -1,7 +1,7 @@
 use crate::api::asset::AssetSignature;
 use crate::commands::edge_app::instance_manifest::InstanceManifest;
 use crate::commands::edge_app::manifest::EdgeAppManifest;
-use crate::commands::edge_app::setting::{Setting, SettingType};
+use crate::api::edge_app::setting::{Setting, SettingType};
 use crate::commands::CommandError;
 use crate::signature::{generate_signature, sig_to_hex};
 use log::debug;
@@ -315,7 +315,7 @@ mod tests {
     use crate::commands::edge_app::instance_manifest::INSTANCE_MANIFEST_VERSION;
     use crate::commands::edge_app::manifest::{Auth, Entrypoint, EntrypointType, MANIFEST_VERSION};
     use crate::commands::edge_app::manifest_auth::AuthType;
-    use crate::commands::edge_app::setting::SettingType;
+    use crate::api::edge_app::setting::{Setting, SettingType};
     use std::fs::File;
     use std::io::Write;
     use temp_env;
